FE-4 Поправил структуру проекта
