+++ web/components/MailInput/MailInput.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import { Form, Input, Select } from 'antd';
 
 const { Option } = Select;
