+++ web/components/TabPaneItem/TabPaneItem.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import useWindowSize from '../../scripts/hooks/useWindowSize';
 import styles from './TabPaneItem.module.scss';
