--- web/pages/editor/index.jsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-export default function Editor() {
-	return (
-		<span>Placeholder</span>
-	);
-}
