+++ web/pages/login/index.jsx
@@ -1,12 +1,13 @@
-import * as React from 'react';
+import React from 'react';
 import { Button, ConfigProvider, Form, Input, Layout } from 'antd';
-const { Content } = Layout;
 import { useRouter } from 'next/router';
 import Header from '../../components/Header/Header';
 import Footer from '../../components/Footer/Footer';
 import styles from '../../public/styles/pages/Login.module.scss';
 import { useState } from 'react';
 import timeout from '../../scripts/timeout';
+
+const { Content } = Layout;
 
 export default function Login() {
 	const [isLoading, setLoading] = useState(false);
