+++ web/public/styles/pages/editor.module.scss
@@ -1,0 +1,5 @@
+.card {
+  flex: 1 0 300px;
+  text-align: center;
+  cursor: pointer;
+}
